boss health sync
sorry for fucking this up

+++ Unity/CleanupCrewV2/CleanupCrewV2/Assets/Scripts/Boss stuff/BossBase.cs
@@ -38,5 +38,5 @@ public abstract class BossBase : MonoBehaviour {
     public abstract void Die();
 
     public float hitpointsForHud
-    { get { return 1 / maxhitpoints * hitpoints; } }
+    { get { return (float) 1 / (float)maxhitpoints * (float)hitpoints; } }
 }
